refactor hash calculation code
